Css adjustment + project page
Adjusting some css content.
+
Creating the projectContentPage.js to render the project content.
Adapting the existing files to work with a new url -> /project/<str:id>,  because  the website works based on the active navItem, when the user access a project page the "Projects" item on Navbar will keep active, if the user clicks on it the projects thumbnails will render