Fixing some erros
On last commit I changed the possibility from user to choose the sizer of the image in pixels to change 3 different sizes, small, medium and big. For a better UX and making easier to keep it responsive.
Here I just fixed the project page display order.  and fixed the way the js pass the data to django